Se arreglan las validaciones, se crea el controlador de Citas (Presenta errores)

+++ Models/Categoria.cs
@@ -10,7 +10,7 @@ namespace GYMIMFC.Models
     {
         [Key]
 
-        public virtual short idCategoria { get; set; }
+        public virtual int idCategoria { get; set; }
         [Display(Name = "Nombre: ")]
         [Required]
         public virtual string NombreCategoria { get; set; }
